Refactor `simple-git`, `parsers/default`, and `ui` modules
Refactor several modules for cleaner code structure and easier maintenance. In `simple-git`, replace configurations loaded in runtime with imports from constants. Relocated `loadConfig` within `utils`. Within `parsers/default`, import functions have been relocated, each function is imported from its own path. Same changes applied in `ui`, where the paths for `COMMIT_PROMPT` and `validatePromptTemplate` have been updated.
