+++ src/lib/ui/editPrompt.ts
@@ -1,7 +1,7 @@
 import { editor } from '@inquirer/prompts'
 import { GenerateReviewLoopOptions } from './generateAndReviewLoop'
-import { COMMIT_PROMPT } from '../langchain/prompts/commitDefault'
-import { validatePromptTemplate } from '../langchain/utils'
+import { COMMIT_PROMPT } from '../langchain/prompts/commit'
+import { validatePromptTemplate } from '../langchain/utils/validatePromptTemplate'
 
 export async function editPrompt(options: GenerateReviewLoopOptions): Promise<string> {
   return await editor({
